refactor: extract voice/drive_row.rs — honest pub(super) seam (deepening 02)
voice/mod.rs documented its pub(super) helpers as the cross-renderer surface,
but a ~225-line drive-row presentation cluster carried a false seam: marked
pub(super) yet consumed by only status.rs, with three of its six functions
having zero callers outside the file. One adapter is a hypothetical seam, not a
real one — the visibility marker mis-taught the next sub-renderer author into
reusing status-specific code (the drive-age cascade, the absent-reserved rule,
the rotation-forecast wording).

Move the cluster (aggregate_drive_info, unmounted_drive_label,
offsite_drive_label + the three private helpers and DriveAggregate) into a
dedicated voice/drive_row.rs owned by its sole consumer, status.rs. The three
helpers with no cross-renderer caller regain private scope; the three entries
stay pub(super). DriveAggregate's fields become pub(super) (the struct moved
from descendant-visible to sibling). mod.rs's remaining pub(super) surface now
reliably means "shared across renderers."

Also relocate the status renderer's tests (everything calling render_status /
render_default_status / render_first_time, plus the status-only posture()
helper) from mod.rs into a new test module in status.rs, so they live next to
the code they exercise. Non-status renderer tests stay in mod.rs (deferred to a
per-renderer follow-up series). Drive-row cluster tests move into drive_row.rs.

No behavior change: every rendered line is pinned by voice_contract.rs (green and
untouched) and the in-file render tests. mod.rs shrinks 6301 -> 4733 lines.

Modules: voice
